requirements.txt
# QuantX Data Builder - Root Requirements
# For local development and testing

# CLI framework
click==8.1.7

# Azure dependencies
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-functions==1.12.0

# Data processing
# yfinance==0.2.36  # Replaced with Tiingo API
pandas==2.2.0
pyarrow==15.0.0

# Statistical analysis
statsmodels==0.14.1
scipy==1.12.0

# Optimization
cvxpy==1.4.2

# Configuration
pyyaml==6.0.1

# Tiingo API client
tiingo==0.16.1

# HTTP requests
requests==2.31.0

# Retry logic
tenacity==8.2.3

# Excel support (for HKEX data)
openpyxl==3.1.2
lxml==5.1.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
mypy==1.7.1
black==23.12.1
python-dotenv==1.0.1
